external monitor and move tests
