+++ modules/api/src/main/scala/vinyldns/api/engine/sqs/SqsConnection.scala
@@ -25,7 +25,7 @@ import com.amazonaws.services.sqs.{AmazonSQSAsync, AmazonSQSAsyncClientBuilder}
 import com.amazonaws.{AmazonWebServiceRequest, AmazonWebServiceResult}
 import com.typesafe.config.Config
 import vinyldns.api.VinylDNSConfig
-import vinyldns.api.route.Monitor
+import vinyldns.core.route.Monitor
 
 object SqsConnection {
